refactor: detail 카드 컴포넌트의 수정삭제 dialog 를 컴포넌트로 분리
